requirements.txt
Django==4.1.5
djangorestframework==3.14.0
django-environ==0.9.0
psycopg2-binary==2.9.5
django-filter==22.1
drf-spectacular==0.25.1
